tests/blueprints: Rename `file` helper to `fixture`
