+++ node-tests/blueprints/helper-addon-test.js
@@ -8,7 +8,7 @@ const emberGenerateDestroy = blueprintHelpers.emberGenerateDestroy;
 const chai = require('ember-cli-blueprint-test-helpers/chai');
 const expect = chai.expect;
 
-const file = require('../helpers/file');
+const fixture = require('../helpers/fixture');
 
 describe('Blueprint: helper-addon', function() {
   setupTestHooks(this);
@@ -21,7 +21,7 @@ describe('Blueprint: helper-addon', function() {
     it('helper-addon foo/bar-baz', function() {
       return emberGenerateDestroy(['helper-addon', 'foo/bar-baz'], _file => {
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper-addon.js'));
+          .to.equal(fixture('helper-addon.js'));
       });
     });
   });
