+++ node-tests/blueprints/helper-test.js
@@ -9,7 +9,7 @@ const setupPodConfig = blueprintHelpers.setupPodConfig;
 const chai = require('ember-cli-blueprint-test-helpers/chai');
 const expect = chai.expect;
 
-const file = require('../helpers/file');
+const fixture = require('../helpers/fixture');
 
 describe('Blueprint: helper', function() {
   setupTestHooks(this);
@@ -22,36 +22,36 @@ describe('Blueprint: helper', function() {
     it('helper foo/bar-baz', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz'], _file => {
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
 
     it('helper foo/bar-baz unit', function() {
       return emberGenerateDestroy(['helper', '--test-type=unit', 'foo/bar-baz'], _file => {
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('tests/unit/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/unit.js'));
+          .to.equal(fixture('helper-test/unit.js'));
       });
     });
 
     it('helper foo/bar-baz --pod', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz', '--pod'], _file => {
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
 
     it('helper foo/bar-baz --pod', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz', '--pod'], _file => {
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
 
@@ -63,18 +63,18 @@ describe('Blueprint: helper', function() {
       it('helper foo/bar-baz --pod', function() {
         return emberGenerateDestroy(['helper', 'foo/bar-baz', '--pod'], _file => {
           expect(_file('app/helpers/foo/bar-baz.js'))
-            .to.equal(file('helper.js'));
+            .to.equal(fixture('helper.js'));
           expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-            .to.equal(file('helper-test/integration.js'));
+            .to.equal(fixture('helper-test/integration.js'));
         });
       });
 
       it('helper foo/bar-baz --pod', function() {
         return emberGenerateDestroy(['helper', 'foo/bar-baz', '--pod'], _file => {
           expect(_file('app/helpers/foo/bar-baz.js'))
-            .to.equal(file('helper.js'));
+            .to.equal(fixture('helper.js'));
           expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-            .to.equal(file('helper-test/integration.js'));
+            .to.equal(fixture('helper-test/integration.js'));
         });
       });
     });
@@ -88,29 +88,29 @@ describe('Blueprint: helper', function() {
     it('helper foo/bar-baz', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz'], _file => {
         expect(_file('addon/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper-addon.js'));
+          .to.equal(fixture('helper-addon.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
 
     it('helper foo/bar-baz', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz'], _file => {
         expect(_file('addon/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper-addon.js'));
+          .to.equal(fixture('helper-addon.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
 
     it('helper foo/bar-baz --dummy', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz', '--dummy'], _file => {
         expect(_file('tests/dummy/app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('app/helpers/foo/bar-baz.js'))
           .to.not.exist;
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
@@ -127,11 +127,11 @@ describe('Blueprint: helper', function() {
     it('helper foo/bar-baz --in-repo-addon=my-addon', function() {
       return emberGenerateDestroy(['helper', 'foo/bar-baz', '--in-repo-addon=my-addon'], _file => {
         expect(_file('lib/my-addon/addon/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper.js'));
+          .to.equal(fixture('helper.js'));
         expect(_file('lib/my-addon/app/helpers/foo/bar-baz.js'))
-          .to.equal(file('helper-addon.js'));
+          .to.equal(fixture('helper-addon.js'));
         expect(_file('tests/integration/helpers/foo/bar-baz-test.js'))
-          .to.equal(file('helper-test/integration.js'));
+          .to.equal(fixture('helper-test/integration.js'));
       });
     });
   });
